move followed hashtag screen to separate package
